% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bateman.R
\name{compute_Bateman}
\alias{compute_Bateman}
\title{Compute the Bateman gradient}
\usage{
compute_Bateman(mating_success, reproductive_success)
}
\arguments{
\item{mating_success}{A vector of mating success outcomes.}

\item{reproductive_success}{A vector of reproductive success outcomes.}
}
\value{
The Bateman gradient.
}
\description{
This function computes the Bateman gradient. The Bateman gradient is the
slope of the regression line of the reproductive success on the mating
success. Following Jones 2009 (and unlike some other work), we consider the
relative reproductive success and the relative mating success to compute the
gradient; i.e. we divide each variable by the population mean before
fitting the linear model.
}
\examples{
compute_Bateman(mating_success = males$Mat_succ,
                reproductive_success = males$Rep_succ)

}
\references{
Jones A.G (2009).On the opportunity for sexual selection, the
Bateman gradient and the maximum intensity of sexual selection. Evolution,
63(7), 1673-1684.
}
\seealso{
\code{\link{test_NonacsB}}
}
