% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{compute_correlation}
\alias{compute_correlation}
\title{Compute a Spearman correlation test between two variables}
\usage{
compute_correlation(var1, var2)
}
\arguments{
\item{var1}{A vector.}

\item{var2}{A vector.}
}
\value{
A vector containing the correlation between two variables, the 
p-value of the correlation test, and the sample size after discarding the 
missing values.
}
\description{
This function computes the Spearman correlation between two variables. It is
a simple wrapper around the R function \code{\link{cor.test}}, with better
display of the output.
}
\examples{
compute_correlation(var1 = males$Mat_succ, var2 = males$Rep_succ)

}
