% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{compute_correlation_table}
\alias{compute_correlation_table}
\title{Compute a Spearman correlation tests for a given cohort and fitness measurement}
\usage{
compute_correlation_table(cohort = NULL, fitness = NULL,
  method = "holm", data = NULL)
}
\arguments{
\item{cohort}{The cohort of males ('C1' or 'C2').}

\item{fitness}{The column name for the fitness component ('Mat_succ' or 'Rep_succ').}

\item{method}{The method for the multiple testing correction, see \code{\link{p.adjust}}.}

\item{data}{The dataset.}
}
\value{
A data.frame containing the predictor, the sample size after
  discarding the missing values, the correlation between two variables, the
  p-value of the correlation test, the p-value after correction for
  multiple testing, and significance stars for the raw and adjusted p-values.
}
\description{
This function computes the Spearman correlation tests between a fitness
component and all predictors. It is a simple wrapper around the R function
\code{\link{cor.test}}, with better display of the output that includes
correction for multiple testing (default = 'holm').
}
\examples{
compute_correlation_table(cohort = 'C1', fitness = 'Mat_succ', method = 'holm', data = males)

}
