% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nonacs.R
\name{compute_Nonacs}
\alias{compute_Nonacs}
\alias{Nonacs}
\title{Compute the the binomial skew index (B) of Nonacs}
\usage{
compute_Nonacs(benef, time)
}
\arguments{
\item{benef}{The vector of benefits (i.e. mating or reproductive success).}

\item{time}{The vector of time-in (i.e. the time each individual spent in the
group).}
}
\value{
The observed Nonacs' binomial skew index value
}
\description{
This function computes the Nonacs' binomial skew index. Nonacs defines it as
an index ``based on the observed variance in a group corrected by the
expected variance if each member had an equal probability of gaining any
given group benefit or reproductive opportunity.''
}
\examples{
compute_Nonacs(benef = c(1, 1, 5), time = c(1, 1, 1))

}
\references{
Nonacs, P. (2000). Measuring and using skew in the study of
social behavior and evolution. The American Naturalist, 156(6), 577-589.
}
\seealso{
\code{\link{test_Nonacs}}
}
