% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{compute_pca}
\alias{compute_pca}
\title{Compute a PCA on horn characteristics}
\usage{
compute_pca(data, digits = 3L)
}
\arguments{
\item{data}{A dataframe used as an input for the PCA.}

\item{digits}{The number of digits for the display.}
}
\value{
A list providing the original data with the first 2 principal
  components added to it, the proportion of variance explained by each
  component, the cummulated proportion of variance explained by each
  component, and the PCA object as returned by \code{\link{prcomp}}.
}
\description{
This function computes a Principal Component Analysis (PCA) on male horn
measurements using \code{\link{prcomp}}. The PCA is used on the centered and
scaled variables.
}
\examples{
pca_C1_males <- compute_pca(males[males$Cohort == 'C1', ])
pca_C2_males <- compute_pca(males[males$Cohort == 'C2', ])

}
\seealso{
\code{\link{plot_pca}} \code{\link{figure_pca}}
}
