% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d_sim}
\alias{d_sim}
\title{d_sim}
\format{A \var{dataframe}.}
\description{
This dataset was used to infer the best measure of relatedness. It contains 600
simulated observations and 12 variables.
}
\examples{
d_sim
}
\references{
This study.
}
\keyword{datasets}
