% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{figure_correlations}
\alias{figure_correlations}
\title{Create the figure showing the correlation between mating/reproductive success and their correlates}
\usage{
figure_correlations(data, savePDF = FALSE)
}
\arguments{
\item{data}{The dataset on which to perform the PCAs.}

\item{savePDF}{A boolean indicating whether to create a PDF of the plot in the
current working directory (TRUE) or not (FALSE, default).}
}
\description{
This function creates the figure showing the correlations. It is a wrapper
around the function \code{\link{plot_correlation}}.
}
\examples{
figure_correlations(data = males, savePDF = TRUE)

}
\seealso{
\code{\link{plot_correlation}}
}
