% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nonacs.R
\name{figure_NonacsB}
\alias{figure_NonacsB}
\title{Create the figure showing the results of the Nonacs' binomial skew analysis}
\usage{
figure_NonacsB(data_males, data_females)
}
\arguments{
\item{data_males}{The dataset for males.}

\item{data_females}{The dataset for females.}
}
\description{
This function creates the figure showing the results of the Nonacs' binomial skew analysis for both sexes.
It is a wrapper around the function \code{\link{plot_NonacsB}}.
}
\examples{
figure_NonacsB(data_males = males, data_females = females)

}
