% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{figure_PCA}
\alias{figure_PCA}
\title{Create the figure showing the PCA results}
\usage{
figure_PCA(data)
}
\arguments{
\item{data}{A dataframe containing the data to be processed.}
}
\description{
This function creates the figure showing the results of the PCAs. It is a wrapper
around the function \code{\link{compute_PCA}} and \code{\link{plot_PCA}}.
}
\examples{
figure_PCA(data = males)

}
\seealso{
\code{\link{compute_PCA}} \code{\link{plot_PCA}}
}
