% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_simulation.R
\name{figure_relatedness_simulation}
\alias{figure_relatedness_simulation}
\title{Plot the results of the simulated relatedness}
\usage{
figure_relatedness_simulation(d_sim)
}
\arguments{
\item{d_sim}{A dataframe of simulated relatedness}
}
\value{
A ggplot object
}
\description{
This function creates a plot of the simulated relatedness
}
\examples{
data(d_sim)
figure_relatedness_simulation(d_sim)

}
