% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testo.R
\name{figure_testosterone}
\alias{figure_testosterone}
\title{Create the figure showing the distribution of testosterone}
\usage{
figure_testosterone(data, savePDF = FALSE)
}
\arguments{
\item{data}{The dataset on which to perform the PCAs.}

\item{savePDF}{A boolean indicating whether to create a PDF of the plot in the
current working directory (TRUE) or not (FALSE, default).}
}
\description{
This function creates the figure showing the distribution of testosterone for
the two cohorts of males and for females. It is a wrapper around the function
\code{\link{plot_testosterone}}.
}
\examples{
figure_testosterone(data = males)

}
\seealso{
\code{\link{plot_testosterone}}
}
