% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{males}
\alias{males}
\alias{females}
\alias{males_agg}
\alias{females_agg}
\title{The datasets}
\format{A \var{dataframe}.}
\description{
Those 4 datasets are those used in the paper. There are two datasets
providing individual-level data (one for males, one for females) and two
datasets providing data aggregated by 'mating success - reproductive success
- male cohort' combination (also one for males and one for females). The
aggregated data are used for ploting only.
}
\examples{
males
females
males_agg
females_agg
}
\references{
This study.
}
\keyword{datasets}
