% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bateman.R
\name{plot_bateman}
\alias{plot_bateman}
\title{Plot the relationship between mating and reproductive success}
\usage{
plot_bateman(data_agg)
}
\arguments{
\item{data_agg}{A dataframe of aggregated data.}
}
\value{
A ggplot object.
}
\description{
This function creates a plot of the relationship between mating and 
reproductive success.
}
\examples{
plot_bateman(data_agg = males_agg)

}
