% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{plot_correlation}
\alias{plot_correlation}
\title{Plot the correlation between mating/reproductive success and their correlates}
\usage{
plot_correlation(data, x, y, xlab = "x-axis title",
  ylab = "y-axis title", limits = NULL)
}
\arguments{
\item{data}{A dataframe containing the data to be processed.}

\item{x}{The name of the x covariate.}

\item{y}{The name of the y covariate.}

\item{xlab}{A title for the x-axis.}

\item{ylab}{A title for the y-axis.}

\item{limits}{The limit for the y-axis of the plot.}
}
\value{
A ggplot object.
}
\description{
This function creates a plot of the correlations using the
package ggplot2.
}
\examples{
plot_correlation(data = males,
                 x = 'Ter_map', y = 'Mat_succ',
                 xlab = 'x-axis title', ylab = 'y-axis title')

}
\seealso{
\code{\link{figure_correlations}}
}
