% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nonacs.R
\name{plot_NonacsB}
\alias{plot_NonacsB}
\title{Plot the result of the null hypothesis testing of the Nonacs' binomial skew
index}
\usage{
plot_NonacsB(x, limits = c(-0.04, 0.08))
}
\arguments{
\item{x}{The object returned by the function \code{\link{test_NonacsB}}}

\item{limits}{}
}
\value{
A ggplot object.
}
\description{
This function creates a plot of the outcome of the null hypothesis testing of
the Nonacs' binomial skew index using the package ggplot2.
}
\examples{
malesRep <- test_NonacsB(benef = males$Rep_succ, time = males$Time, keep_H0 = TRUE)
plot_NonacsB(x = malesRep)

}
