% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatedness.R
\name{plot_relatedness}
\alias{plot_relatedness}
\title{Plot the distribution of relatedness}
\usage{
plot_relatedness(relatedness, ID, limits = c(-0.25, 0.5))
}
\arguments{
\item{relatedness}{}

\item{ID}{}
}
\value{
A ggplot object.
}
\description{
This function creates a plot of the distribution of relatedness using the
package ggplot2.
}
\examples{
plot_relatedness(relatedness = males$Related[males$Cohort == "C1"], ID = males$No[males$Cohort == "C1"])

}
\seealso{
\code{\link{figure_relatedness}}
}
