% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testo.R
\name{plot_testosterone}
\alias{plot_testosterone}
\title{Plot the distribution of testosterone}
\usage{
plot_testosterone(data, limits = c(0, 120))
}
\arguments{
\item{data}{A dataframe containing the data to be processed.}

\item{limits}{The limit for the y-axis of the plot.}
}
\value{
A ggplot object.
}
\description{
This function creates a plot of the distribution of testosterone using the
package ggplot2.
}
\examples{
plot_testosterone(data = males[males$Cohort == 'C1', ])

}
\seealso{
\code{\link{figure_testosterone}}
}
