% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nonacs.R
\name{test_NonacsB}
\alias{test_NonacsB}
\title{Test the significance of the Nonacs' binomial skew index}
\usage{
test_NonacsB(benef, time, nsim = 100000L, keep_H0 = FALSE,
  digits = Inf, seed = 1L)
}
\arguments{
\item{benef}{The vector of benefits (i.e. mating or reproductive success).}

\item{time}{The vector of time-in (i.e. the time each individual spent in the
group).}

\item{nsim}{The number of simulation to run.}

\item{keep_H0}{A boolean indicating whether to export the values of B under
the null hypothesis.}

\item{digits}{The number of digits for the display.}

\item{seed}{The seed for the random number generator.}
}
\value{
A list with the observed Nonacs' binomial skew index value, its
  corresponding p-value and (optionally) the values computed on data
  simulated under the null hypothesis.
}
\description{
This functions compare the observed values of the Nonacs' binomial skew index
to expectation under the null hypothesis of a contribution to the mating or
breeding pool proportional to the time spent in the group.
}
\examples{
test_NonacsB(benef = males$Rep_succ, time = males$Time)

}
\seealso{
\code{\link{compute_NonacsB}}
}
