% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{compute_correlation}
\alias{compute_correlation}
\title{Compute a Spearman correlation test between two variables}
\usage{
compute_correlation(var1, var2)
}
\arguments{
\item{var1}{A vector.}

\item{var2}{A vector.}
}
\value{
A vector containing the correlation between two variables, the
p-value of the correlation test, and the sample size after discarding the
missing values.
}
\description{
This function computes the Spearman correlation between two variables. It is
a simple wrapper around the R function \code{\link{cor.test}}, with better
display of the output.
}
\note{
We set the argument \code{exact} to \var{FALSE} in the call to \code{\link{cor.test}},
so that the presence of ties in some correlation tests would not change
how the p-values are computed (exact tests are not possible in the presence of ties for this implementation of
the test). Otherwise p-values obtained can sometimes be inconsistent across tests (ex: smaller p-values despite higher rho and same sample
size.)
}
\examples{
compute_correlation(var1 = males$Mat_succ, var2 = males$Rep_succ)

}
\seealso{
\code{\link{compute_correlation_table}}
}
