% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{compute_PCA}
\alias{compute_PCA}
\title{Compute a PCA on horn characteristics}
\usage{
compute_PCA(data)
}
\arguments{
\item{data}{A dataframe containing the data to be processed.}
}
\value{
A list providing the original data with the first 2 principal
components added to it, the proportion of variance explained by each
component, the cummulated proportion of variance explained by each
component, and the PCA object as returned by \code{\link{prcomp}}.
}
\description{
This function computes a Principal Component Analysis (PCA) on male horn
measurements using \code{\link{prcomp}}. The PCA is computed on the centered
and scaled variables.
}
\examples{
PCA_C1_males <- compute_PCA(males[males$Cohort == 'C1', ])
PCA_C2_males <- compute_PCA(males[males$Cohort == 'C2', ])

}
\seealso{
\code{\link{plot_PCA}} \code{\link{figure_PCA}}
}
