% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datasets}
\alias{datasets}
\alias{males}
\alias{females}
\alias{rhinos_agg}
\title{The datasets}
\format{A \var{dataframe}.}
\description{
Those 3 datasets are those used in the paper. There are two datasets
providing individual-level data (one for males, one for females) and one
datasets providing data aggregated by 'mating success - reproductive success
- male cohort' combination (one for both sexes). The
aggregated dataset is used for ploting only.
}
\examples{
males
females
rhinos_agg
}
\references{
This study.
}
\keyword{datasets}
