% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bateman.R
\name{figure_bateman}
\alias{figure_bateman}
\title{Create the figure showing the relationship between mating and reproductive success for both sexes}
\usage{
figure_bateman(data_agg, savePDF = FALSE)
}
\arguments{
\item{data_agg}{A dataframe of aggregated data.}

\item{savePDF}{A boolean indicating whether to create a PDF of the plot in the
current working directory (TRUE) or not (FALSE, default).}
}
\description{
This function creates the figure showing the results of the the relationship between mating and reproductive success for both sexes.
It is a wrapper around the function \code{\link{plot_bateman}}.
}
\examples{
figure_bateman(data_agg = rhinos_agg)

}
\seealso{
\code{\link{plot_bateman}}
}
