% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{figure_correlations}
\alias{figure_correlations}
\title{Create the figures showing the correlation between mating/reproductive success and their correlates}
\usage{
figure_correlations(data, which = c("mating", "repro"))
}
\arguments{
\item{data}{A dataframe containing the data to be processed.}

\item{which}{The figure to be drawn ('mating' or 'repro').}
}
\description{
This function creates the figures showing the correlations. It is a
wrapper around the function \code{\link{plot_correlation}}. The function
can either illustrate the correlations with the mating success or those with
the reproductive success depending on the value of the argument \code{which}.
}
\examples{
figure_correlations(data = males, which = 'mating')
figure_correlations(data = males, which = 'repro')

}
\seealso{
\code{\link{plot_correlation}}
}
