% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nonacs.R
\name{figure_NonacsB}
\alias{figure_NonacsB}
\title{Create the figure showing the results of the Nonacs' binomial skew analysis}
\usage{
figure_NonacsB(data_males, data_females, savePDF = FALSE)
}
\arguments{
\item{data_males}{The dataset for males.}

\item{data_females}{The dtaset for females.}

\item{savePDF}{A boolean indicating whether to create a PDF of the plot in the
current working directory (TRUE) or not (FALSE, default).}
}
\description{
Create the figure showing the results of the Nonacs' binomial skew analysis
}
\examples{
figure_NonacsB(data_males = males, data_females = females)

}
