% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{figure_pca}
\alias{figure_pca}
\title{Create the figure showing the PCA results}
\usage{
figure_pca(data, savePDF = FALSE)
}
\arguments{
\item{data}{The dataset on which to perform the PCAs.}

\item{savePDF}{A boolean indicating whether to create a PDF of the plot in the
current working directory (TRUE) or not (FALSE, default).}
}
\description{
This function creates the figure showing the results of the PCAs. It is a wrapper
around the function \code{\link{do_pca}} and \code{\link{plot_pca}}.
}
\examples{
figure_pca(data = males)

}
\seealso{
\code{\link{do_pca}} \code{\link{plot_pca}}
}
