% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{figure_PCA}
\alias{figure_PCA}
\title{Create the figure showing the PCA results}
\usage{
figure_PCA(data, savePDF = FALSE)
}
\arguments{
\item{data}{The dataset on which to perform the PCAs.}

\item{savePDF}{A boolean indicating whether to create a PDF of the plot in the
current working directory (TRUE) or not (FALSE, default).}
}
\description{
This function creates the figure showing the results of the PCAs. It is a wrapper
around the function \code{\link{compute_PCA}} and \code{\link{plot_PCA}}.
}
\examples{
figure_PCA(data = males)

}
\seealso{
\code{\link{compute_PCA}} \code{\link{plot_PCA}}
}
