% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testo.R
\name{figure_testosterone}
\alias{figure_testosterone}
\title{Create the figure showing the distribution of testosterone}
\usage{
figure_testosterone(data)
}
\arguments{
\item{data}{A dataframe containing the data to be processed.}
}
\description{
This function creates the figure showing the distribution of testosterone for
the two cohorts of males. It is a wrapper around the function
\code{\link{plot_testosterone}}.
}
\examples{
figure_testosterone(data = males)

}
\seealso{
\code{\link{plot_testosterone}}
}
