% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matingRhinos-package.R
\docType{package}
\name{matingRhinos-package}
\alias{matingRhinos-package}
\alias{matingRhinos}
\title{Analysis of the Mating, Reproductive Success and Their Correlates in White
Rhinos}
\description{
This R package aims at providing the data and documenting the R code behind
the analysis of the paper entitled "Mate choice, reproductive success and
inbreeding in white rhinos" by Kretzschmar Petra et al. (in prep).
}
\details{
This package has not been conceived for general use!

All main functions of this package contain a small documentation and examples
which could be useful for those who try to understand our code. Type
\code{ls("package:matingRhinos")} for a list of all exported functions and
datasets.

You may directly explore the files contained in the package on GitHub at
\url{https://github.com/courtiol/matingRhinos}, or after uncompressing the
content of the *.tar.gz file (link available on the GitHub as well). You can
use the R function \code{\link{untar}} to extract the content of the tarball.

The package contains all the original data.

In the examples below, we provide the workflow leading the results presented
in the paper.
}
\examples{


################################################
##  Setting general options for this workflow ##
################################################

### Note: set the following options to TRUE or FALSE depending on what you want.

save_figures_on_disk <- TRUE # export figures on drive?


#############################
##  Preparing the datasets ##
#############################

### 1. We split the dataset of all males per cohorts:
malesC1 <- droplevels(males[males$Cohort == "C1", ])
malesC2 <- droplevels(males[males$Cohort == "C2", ])

### 2. We visuallise the first 6 rows of all the datasets:
head(males)
head(malesC1)
head(malesC2)
head(females)


#########################
##  Parentage analysis ##
#########################

### 1. Note: the bioinformatic work leading to parentage has not been done using R.

### 2. Result about the higher relatedness among C2 males than among C1 males:#' 
round(mean(malesC1$Related_mean), digits = 3L)
round(sd(malesC1$Related_mean), digits = 3L)
round(mean(malesC2$Related_mean), digits = 3L)
round(sd(malesC2$Related_mean), digits = 3L)

wilcox.test(malesC1$Related_mean, malesC2$Related_mean)


######################################
##  Mating and reproductive success ##
######################################

### 1. Computing the skewness tests:#' 
test_NonacsB(benef = males$Rep_succ, time = males$Time, digits = 3L)
test_NonacsB(benef = males$Mat_succ, time = males$Time, digits = 3L)
test_NonacsB(benef = females$Rep_succ, time = females$Time, digits = 3L)


### 2. Creating figure 1:#' 
figure_NonacsB(data_males = males, data_females = females, savePDF = save_figures_on_disk)

### 3. Relationship between mating and reproductive success: 
cor.test(~ Mat_succ + Rep_succ, data = malesC1, method = "spearman")

cor.test(~ Mat_succ + Rep_succ, data = malesC2, method = "spearman")

 
### 2. Creating the figure pannel:

figure_Bateman(data_agg = rhinos_agg, savePDF = save_figures_on_disk)



#############################################################################
##  Horn characteristics as a correlate of mating and reproductive success ##
#############################################################################

### 1. Running the Principal Component Analyses:

pca_C1_males <- do_pca(data = malesC1)
pca_C2_males <- do_pca(data = malesC2)


### 2. Computing the correlations and testing them:

cor.test(~ PC1 + Mat_succ, data = pca_C1_males$data, method = "spearman")
cor.test(~ PC1 + Rep_succ, data = pca_C1_males$data, method = "spearman")

cor.test(~ PC1 + Mat_succ, data = pca_C2_males$data, method = "spearman")
cor.test(~ PC1 + Rep_succ, data = pca_C2_males$data, method = "spearman")


### 3. Creating the figure pannel:

figure_pca(data = males, savePDF = save_figures_on_disk)



#####################################################################
##  Testosterone as a correlate of mating and reproductive success ##
#####################################################################

### 1. Computing the correlations and testing them:

cor.test(~ Testo_mean + Mat_succ, data = malesC1, method = "spearman")
cor.test(~ Testo_mean + Rep_succ, data = malesC1, method = "spearman")
cor.test(~ Testo_mean + Mat_succ, data = malesC2, method = "spearman")


### 2. Creating the figure pannel:

figure_testosterone(data = males, savePDF = save_figures_on_disk)



####################################################################
##  Relatedness as a correlate of mating and reproductive success ##
####################################################################

### 1. Computing the correlations and testing them:

cor.test(~ Related_mean + Mat_succ, data = malesC1, method = "spearman")
cor.test(~ Related_mean + Rep_succ, data = malesC1, method = "spearman")
cor.test(~ Related_mean + Mat_succ, data = malesC2, method = "spearman")


### 2. Creating the figure pannel:

figure_correlations(data = males, savePDF = save_figures_on_disk)



}
\references{
Kretzschmar Petra et al. (in prep)
Mate choice, reproductive success and inbreeding in white rhinos.
}
\keyword{package}
