% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bateman.R
\name{plot_Bateman}
\alias{plot_Bateman}
\title{Plot the relationship between mating and reproductive success}
\usage{
plot_Bateman(data_agg, sex)
}
\arguments{
\item{data_agg}{A dataframe of aggregated data.}

\item{sex}{The sex of the rhinos to plot ('males' or 'females').}
}
\value{
A ggplot object.
}
\description{
This function creates a plot of the relationship between mating and 
reproductive success.
}
\examples{
plot_Bateman(data_agg = rhinos_agg, sex = 'males')
plot_Bateman(data_agg = rhinos_agg, sex = 'females')

}
\seealso{
\code{\link{figure_Bateman}}
}
