% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{plot_pca}
\alias{plot_pca}
\title{Plot the result of a PCA}
\usage{
plot_pca(pca)
}
\arguments{
\item{pca}{The object returned by the function \code{\link{do_pca}}.}
}
\value{
A ggplot object.
}
\description{
This function creates a plot of the outcome of a Principal Component Analysis
(PCA) using the package ggplot2.
}
\examples{
pca_C1_males <- do_pca(males[males$Cohort == "C1", ])
pca_C2_males <- do_pca(males[males$Cohort == "C2", ])
plot_pca(pca_C1_males)
plot_pca(pca_C2_males)

}
\seealso{
\code{\link{do_pca}} \code{\link{figure_pca}}
}
