% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{plot_PCA}
\alias{plot_PCA}
\title{Plot the result of a PCA}
\usage{
plot_PCA(x)
}
\arguments{
\item{x}{The object returned by the function \code{\link{compute_PCA}}.}
}
\value{
A ggplot object.
}
\description{
This function creates a plot of the outcome of a Principal Component Analysis
(PCA) using the package ggplot2.
}
\examples{
PCA_C1_males <- compute_PCA(males[males$Cohort == 'C1', ])
PCA_C2_males <- compute_PCA(males[males$Cohort == 'C2', ])
plot_PCA(PCA_C1_males)
plot_PCA(PCA_C2_males)

}
\seealso{
\code{\link{compute_PCA}} \code{\link{figure_PCA}}
}
