% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatedness.R
\name{plot_relatedness}
\alias{plot_relatedness}
\title{Plot the distribution of relatedness}
\usage{
plot_relatedness(data, limits = c(0, 0.5), to = "males")
}
\arguments{
\item{data}{A dataframe containing the data to be processed.}

\item{limits}{The limit for the y-axis of the plot.}

\item{to}{Either 'males' or 'females', to indicate of the relatedness between
the focal male is computed to males or females.}
}
\value{
A ggplot object.
}
\description{
This function creates a plot of the distribution of relatedness using the
package ggplot2.
}
\examples{
plot_relatedness(data = males, to = 'males')
plot_relatedness(data = males, to = 'females')

}
\seealso{
\code{\link{figure_relatedness}}
}
